fix nested sync calls test
