+++ tests/core/utilities/test_synchronize.py
@@ -1,4 +1,5 @@
 import asyncio
+import pytest
 
 from web3._utils.async_tools import (
     sync,
@@ -17,3 +18,13 @@ def test_sync_coro():
 
 def test_sync_awaitable():
     assert sync(coro_sleeper()) == "Im awake!"
+
+
+async def nested_sync():
+    await asyncio.sleep(0)
+    return sync(asyncio.sleep(0))
+
+
+@pytest.mark.asyncio
+async def test_async_calling_sync():
+    sync(nested_sync)
